lakeshore425: Move lakeshore425 agent into package
