+++ docker-compose.yml
@@ -30,13 +30,6 @@ services:
       - "socs"
 
   # --------------------------------------------------------------------------
-  # The Lakeshore 425 Agent
-  # --------------------------------------------------------------------------
-  ocs-lakeshore425-agent:
-    image: "ocs-lakeshore425-agent"
-    build: ./agents/lakeshore425/
-
-  # --------------------------------------------------------------------------
   # The Pysmurf Controller Agent
   # --------------------------------------------------------------------------
   ocs-pysmurf-agent:
