+++ socs/agents/ocs_plugin_so.py
@@ -15,6 +15,7 @@ for n, f in [
         ('FTSAerotechAgent', 'fts_aerotech/agent.py'),
         ('Lakeshore240Agent', 'lakeshore240/agent.py'),
         ('Lakeshore372Agent', 'lakeshore372/agent.py'),
+        ('Lakeshore425Agent', 'lakeshore425/agent.py'),
         ('RotationAgent', 'hwp_rotation/agent.py'),
 ]:
     ocs.site_config.register_agent_class(n, os.path.join(root, f))
